TicketDiary006 Adds middleware for displaying database results
