+++ ToDo.md
@@ -1,14 +1,15 @@
 # Questions
 
-Do I need to test each get route?
-Do I need to test the index.js file?
-What's the difference between app.use and router.use?
-When would you use a default export and a named export?
+- Do I need to test each get route?
+- Do I need to test the index.js file?
+- What's the difference between app.use and router.use?
+- When would you use a default export and a named export?
 
 # _To Do_
 
-- [] Create a databaseConnection file and test 005
 - [] Create a middleware which connects to database 006
 - [] Add error handling 007
+- [] Add logging 008
 
+- [x] Create a databaseConnection file and test 005
 - [x] Add set engine stuff 004
