+++ package.json
@@ -5,7 +5,7 @@
   "main": "index.js",
   "scripts": {
     "test": "jest",
-    "start": "node -r esm src/script.js",
+    "start": "node -r esm src/index.js",
     "dev": "nodemon -r esm src/index.js"
   },
   "repository": {
